// apps/standard-reader/src/lib/leaflet/types.ts
export const LEAFLET_BLOCK = {
  text: "pub.leaflet.blocks.text",
  header: "pub.leaflet.blocks.header",
  blockquote: "pub.leaflet.blocks.blockquote",
  horizontalRule: "pub.leaflet.blocks.horizontalRule",
  unorderedList: "pub.leaflet.blocks.unorderedList",
  unorderedListItem: "pub.leaflet.blocks.unorderedList#listItem",
  orderedList: "pub.leaflet.blocks.orderedList",
  orderedListItem: "pub.leaflet.blocks.orderedList#listItem",
  bskyPost: "pub.leaflet.blocks.bskyPost",
  image: "pub.leaflet.blocks.image",
  code: "pub.leaflet.blocks.code",
  html: "pub.leaflet.blocks.html",
  iframe: "pub.leaflet.blocks.iframe",
  website: "pub.leaflet.blocks.website",
  math: "pub.leaflet.blocks.math",
  button: "pub.leaflet.blocks.button",
  poll: "pub.leaflet.blocks.poll",
  page: "pub.leaflet.blocks.page",
  separator: "pub.leaflet.blocks.separator",
  standardSitePost: "pub.leaflet.blocks.standardSitePost",
  standardSitePublication: "pub.leaflet.blocks.standardSitePublication",
  imageGallery: "pub.leaflet.blocks.imageGallery",
  imageGalleryImage: "pub.leaflet.blocks.imageGallery#image",
  signup: "pub.leaflet.blocks.signup",
} as const;

export const LEAFLET_PAGE = {
  linearDocument: "pub.leaflet.pages.linearDocument",
  linearDocumentBlock: "pub.leaflet.pages.linearDocument#block",
  canvas: "pub.leaflet.pages.canvas",
  canvasBlock: "pub.leaflet.pages.canvas#block",
} as const;

export const LEAFLET_CONTENT = "pub.leaflet.content";